Set task and activity types when adding to task.
Activities from the home package were causing tasks types to change from
application to home. This was not the intention of setting the task type
when adding an activity. This change sets the task type to the inherent
type of the first activity added. All subsequent activities added to the
task then inherent the task's type overriding the inherent type of the
task.

Fixes bug 9972495.

Change-Id: Ib77675aea790ea64d4f166af62c7138e89356c13