Revert "Revert "Take the input device into account for meta state""
This reverts commit d2e52c8119e2763e1eb6458067dd1075cd400d51, which
is equivalent to re-applying 7dfddd8aabf07f634f6d810859c16ea5d4040967.

The change ("Take the input device into account for meta state") was
valid but caused CTS test failures because the test was wrong. The
test injected a key event with the meta state missing. This faulty
test was masked by the old toggle behavior.