Allow next app transition to change when screen is off.
Prior to this change any attempts to set the app transition when the
screen was off would fail. That would cause the
SOFT_INPUT_IS_FORWARD_NAVIGATION flag to be incorrect when launching
an activity such as the alarm before the screen turns on.

Without SOFT_INPUT_IS_FORWARD_NAVIGATION in the correct state the
android:windowSoftInputMode="stateHidden" attribute would be
mishandled leaving an IME on the screen along with the alarm activity.

Fixes bug 10817166.

Change-Id: I8ad97e67ebab2069e753e957b5d253e36dcdd3e5