Fix for jank when fast-scrolling DeskClock's world cities list in RTL
This is a fix for bug 10863979.

The onPreDraw() method in TextView was canceling the draw every time a
the horizontal scroll changed. These dropped frames were visible as
jank. The fix is to never cancel the draw. It seems likely that the
previous rationale for canceling draw is no longer valid, due to other
fixes, because the scroll state will have been fully updated.

Change-Id: I8c2da5450723f993055a49b1cb57320c08f2a11e