Decouple construction of WebView, CookieSyncManager, WebViewDatabase
Bug 10932261

Most of this flow exists purely to get the Context from
CookieSyncManager.createInstance over to WebViewDatabaseClassic. Make
that depenency more explicit (with a TODO to remove it) and this allows
a much simpler CookieSyncManager implementation for the normal case.
Note after this patch, CookieSyncManager.getInstance() is technically fine
to call without a prior call to createInstance, but retaining the
ordering requirement as a convenience for anyone developing on new OS
but still supporting the older versions.
(Note that CookieSyncManager instance is not required for correct
operation of either existing webview, so these changes only impact
the public API contract of object lifetimes, not the underlying
implementation)

Change-Id: I51fdd6622704f1c749277fee6df2f84ac4c704d2